adds redirects to clientside
